backend/requirements.txt
# requirements.txt

Flask
Flask-CORS
google-generativeai
python-dotenv
PyPDF2